Update patch set 2
Patch Set 2:

> (2 comments)
 > 
 > getfattr would be nice
 > 
 > even nicer though would be to figure out why `unsquashfs -x` isn't
 > working.  the commit message doesn't seem like it's been fully
 > triaged.

is there a way to get the version of unsquashfs that's used in the signer? the version that's available in Goobuntu Trusty (unsquashfs version 4.2-git (2013/03/13)) seems to be too old, and is reproducing the issue I'm seeing.

and not that i just mentioned this, it looks like the one in the chroot is newer (unsquashfs version 4.3 (2014/05/12)). lemme try with that one. would you happen to know if this script is run in a chrome os chroot or outside of one?

Patch-set: 2