Fix test failures when TPM2_MODE is defined
An earlier commit had added this:

  ifeq (${TPM2_MODE},)
  # TODO(apronin): tests for TPM2 case?
  TEST_NAMES += \
          tests/tlcl_tests \
          tests/rollback_index2_tests
  endif

but left this:

  .PHONY: runmisctests
  runmisctests: test_setup
          ${RUNTEST} ${BUILD_RUN}/tests/rollback_index2_tests
          [...]
          ${RUNTEST} ${BUILD_RUN}/tests/tlcl_tests

So if TPM2_MODE is not null, those two test targets won't be
built. This CL puts those two into the same guard, so that
they won't be attempted if they're not built.

BUG=chrome-os-partner:57727
BRANCH=all
TEST=manual

Before, this fails:

  FEATURES=test emerge-reef vboot_reference

Now, it passes.

Change-Id: Ic00f9f867d3d9c719d797907f00fda8bc5044504
Signed-off-by: Bill Richardson <wfrichar@chromium.org>