Update patch set 1
Patch Set 1:

It's not permanently set.  It's just sticky until the OS clears it.  The intent was that the {request, done} bits represent a state.

When the OS requests a clear, it sets {request=1, done=0}.

Multiple reboots may occur after that.  

If any of them get as far as the code which handles request, the state may transition to {request=0, done=1} if successful or {request=0, done=0} if failure.

Multiple reboots may occur after that.

Whenever the OS finally boots, it can check for {request=0, done=1} to know if the clear actually took place.

This change removes that robustness against multiple reboots.

If the OS is using some other method to determine that the TPM owner was cleared (say, cryptohome --action=tpm_status), then we should just remove clear_tpm_owner_done entirely as it serves no purpose.

(Just checked: the only reference outside of vboot's own unit tests is src/platform/factory/py/test/pytests/clear_tpm_owner_request.py, and nothing ever calls it with only_check_clear_done=1 - so it never actually looks at it either.  So, let's just nuke it out of crossystem and ToT vboot entirely, and mark the bit as unusable in nvstorage because some platforms shipped with RO firmware that sets it.)

Patch-set: 1
Label: Code-Review=0