runfutiltests: Stop writing stderr to file
Currently, scripts run by run_test_scripts.sh redirect stderr to a
file. This makes debugging less comfortable especially when tests
fail on a build server.

This change removes set -x options from the scripts. It also reduces
outputs to stderr. When errors are expected stderr is redirected
to stdout, which will be recorded in a file.

BUG=none
BRANCH=none
TEST=make runtests -j

Change-Id: I79c4c1a7c8367748cecedfc926016f5a8c69e5b2
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>