Update patch set 2
Patch Set 2:

> if it's scanning the whole device, and there isn't a valid GPT, then why isn't that enough ?  doesn't reading the primary GPT (and failing) work here ?

The problem here is that we don't even want to try looking for a GPT on certain devices (like RPMB), because they're not directly readable like a normal block device and trying to do so causes weird error messages. Even if those are harmless in this case, I think it's better to avoid just throwing open()/read() syscalls at any random device node we find to see what sticks. It's better if we can already figure out by the name (or position in the list) whether it's worth (and safe) to try that device or not.

> so maybe we can just search for gpt partition table on the devices which are actually contains partitions(by checking the /proc/partitions entries).

Yeah, I like that idea actually. But I don't think the code you wrote for now does a great job at it... it can't tell that something doesn't have partitions when it sees it for the first time (e.g. loop-0).

Instead, I think you should always work with two lines at once (e.g. partname[] and partname_next[]). If partname_next[] is EOF, we already know that partname[] can't be a device with partitions. If it isn't, then you know that partname[] is a real device with partitions iff !strncmp(partname, partname_next, strlen(partname)). After that you strcpy(partname, partname_next) and read the next line into partname_next.

Patch-set: 2
Label: Code-Review=0