vboot_ui_menu: Show Developer options on USB failure
Currently, if the boot from USB fails after pressing Ctrl-U or selecting
"Boot USB Image", only a blank screen is shown instead of showing the current
menu. There is no option for the user to do anything else except wait for the
timeout after which boot from fixed disk is attempted. This does not seem like
an intuitive boot flow.
Hence if the USB boot fails display the current menu, so that the user can
attempt something else.

BUG=None
BRANCH=None
TEST= verfied that menu is displayed on USB boot failure from developer screen.

Change-Id: Ide3967be7bba3d87c8a545a0f4ed52da44150fd0
Signed-off-by: Rizwan Qureshi <rizwan.qureshi@intel.com>