crossystem: Remove defunct sw_wpsw_boot field
The sw_wpsw_boot field doesn't work correctly on the vast majority of
platforms (it seems like it only ever worked on BayTrail). It also isn't
used anywhere in the codebase (only other reference is a comment about
how it doesn't work for most platforms in factory_installer.sh), and
it's no longer clear what it was meant for in the first place
(b/35510092 hints at needing it for some planned feature that was never
implemented). Let's get rid of it to avoid confusing people.

If userspace tools need to know the software write-protect state, they
can instead run flashrom directly. For feedback reports, this output is
already included in the "verified boot" section.

BRANCH=none
BUG=chromium:508269,chromium:742685
TEST=none

Change-Id: I8975b1e2c8e604b4cb48d092c13b923b4db2d207
Signed-off-by: Julius Werner <jwerner@chromium.org>