[unibuild] Support for unibuild firmware signing
For design context, see go/cros-unibuild-signing

This adds support for multiple, shared firmware images from a unified
build that needs to be signed with different OEM specific keys.

It uses a signer_config.csv file (that is generated by pack_firmware.py)
to determine which images need to be signed with which keys.

BUG=b:64842314
TEST=TODO
BRANCH=None

Change-Id: Id3711bbe73dfe652184bc046b5f642c30b8d1627