detachables: Disable up/down arrow keys in TO_DEV menu
The ec is untrusted so we need the user to use the volume
buttons in order to confirm that they really want to enable
dev mode.  Thus, disabling up/down arrow keys in TO_DEV
menu.  Will need to add text to this screen to clarify to
the user that they will need to use volume keys instead.
Beep when user tries to use up/down arrow in the TO_DEV
menu so that the user knows they're doing something wrong.

BUG=b:65546569
BRANCH=None
TEST=test out using volume keys in TO_DEV menu and make
     sure that they are disabled.

Change-Id: Ifc7183c7ca35efaf079abb196a90ab7305380642
Signed-off-by: Shelley Chen <shchen@chromium.org>