Makefile: Build futility without utillib to extend arch support
This removes utillib from the futility build and adds only the required
files from utillib to its build. Utillib includes arch-specific
components (required for crossystem) that only have a limited number of
supported architectures, which is undesirable for futility.

The number of archs that futility can run on is no longer restricted.

BUG=None
BRACH=None
TEST=Build futility and futility_s

Change-Id: Ia9ad2a72b374b5cc3d6004d36c0b7090dbcc97a2
Signed-off-by: Paul Kocialkowski <contact@paulk.fr>