vboot: Fix for Power button key press for quick exit in Recovery Insert Screen
The functionality for quick exit in recovery insert screen is not working.
(for users who has entered recovery screen accidently). Shutdown request
return logic was based on current_menu and it was never satisfied as
current_menu is updated with the function call vb2_update_menu() which is
executed before the logic. Now the return logic is modified to
prev_menu which holds the previous screen menu VB_MENU_RECOVERY_INSERT.

BUG=none
BRANCH=none
TEST=verified power button key functionality in Recovery Insert
screen

Change-Id: I495d68befe057dcea26b71a648fb46c9b9d760cc
Signed-off-by: Divya Chellap <divya.chellappa@intel.com>