detachables: Use Volup+Voldown to exit recovery graphic
Previously, we were able to exit the initial recovery graphic with
either the volume up or volume down button.  However, we would only
like the user to exit the screen when they know what they are doing,
we are making it so that they can only intentionally exit (with volup
+ voldown simultaneously pressed).  This also allows us to reclaim the
volume up and volume down keys for language changes in the non-menu
screens in the future.

BUG=b:67371896
BRANCH=None
TEST=reboot into recovery and press volume up + volume down keys.
     Make sure we exit to the recovery menu screen.

Change-Id: I7dfb4a0e42c8b88e9b075e886cea1adfe248246c
Signed-off-by: Shelley Chen <shchen@chromium.org>