Update patch set 1
Patch Set 1:

Sorry, did I miss a discussion somewhere? The last state I thought we were at (from the email thread with Maria) was that we would be doing language cycling directly with the volume buttons and never show any menus on screens that don't have to do with developer mode. On NOGOOD and BROKEN the double-button-press would directly give you debug info and on INSERT it would reveal the secret menu (with options for debug info as well as dev mode transition).

I really preferred that version over the one with a stub menu on all screens. Did we have a reason to change the plans again?

Patch-set: 1
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>