Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Sorry, did I miss a discussion somewhere? The last state I thought we were at (from the email thread with Maria) was that we would be doing language cycling directly with the volume buttons and never show any menus on screens that don't have to do with developer mode. On NOGOOD and BROKEN the double-button-press would directly give you debug info and on INSERT it would reveal the secret menu (with options for debug info as well as dev mode transition).
> 
> I really preferred that version over the one with a stub menu on all screens. Did we have a reason to change the plans again?

Yes, Randall and I actually had a discussion this week about this.  I was having some issues trying to display the debug information if the volume keys were being used a scrolling through the languages as well.  We thought that this was a cleaner and more consistent method of adding that debug and languages functionality to the legacy screens.  Additionally, this would allow us to add any extra functionality to these screens in the future if needed.

Patch-set: 1
Reviewer: Gerrit User 1117900 <1117900@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>