EC-EFS: Enter recovery if update fails too many times
On the system with an EC which has two image slots, it's possible that
all the code to update EC image is successful yet vboot still sees
unexpected hash from an EC upon reboot.

For example, if a bios contains an EC image with a signature signed by
a wrong key, EFS EC falls back to the previous slot. Vboot currently
does not handle this case. Thus, it tries to update the EC with an
incompatible image again.

When this happens, a user sees 'applying critical update' screen
repeatedly. This patch makes vboot detect this update loop and show the
'broken' screen instead.

BUG=b:71719323
BRANCH=none
TEST=Flash EC SPI with an image which has a different RO key. Boot DUT
and let it run software sync. EC-RO rejects the updated image and falls
back to the previous one. The update counter is incremented and vboot
shows the 'broken' screen, requesting recovery.

Change-Id: I725ffe8db05cddb6d2d26942275e61681bcaf4e5
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>