Update patch set 7
Patch Set 7:

This commit is about to be reverted in master, so I don't think it's a good idea to start cherry-picking it for device branches. This commit needs some more work, as it apparently broke automated tests. I'll try to get back to solving this issue when I get the chance.

Patch-set: 7