Update patch set 3
Patch Set 3:

This reboot ec into RO doesn't make a lot of sense to me. One could have easily rebooted again on host side:

if dev_mode and told EC to sysjump to RW:
  reboot

No need to get into the bowels of vboot for that as it could have been handled in depthcharge proper.

Patch-set: 3