Update patch set 4
Patch Set 4:

(2 comments)

I'm not really sure what you're trying to achieve with this patch... "I need the firmware to do something special for display init" _should_ be a solved problem for us. If you select VBOOT_OPROM_MATTERS and correctly guard your DP-connection code with display_init_required() in coreboot, all of this should work the way you need it to.

Maybe there's some edge case with EC-EFS that doesn't yet work right? But in that case we should fix that in the existing ("OPROM") framework rather than reinventing the wheel.

Patch-set: 4