ec_sync: Only do aux fw update if also doing EC update
Aux firmware updates have important caveats that we should be
careful about when we try to apply updates:

1) an update that fails once could continue to fail so the system
could get stuck in an update failure loop.

2) some devices have limited OTP space for updates so a continued
update failure could result in a completely unusable device.

Because of these concerns it seems safer to only attempt to update
the aux firmware when the EC is also being updated.  This way the
aux firmware update will only be attempted once for each overall
firmware update for the device.

Additionally, if the check to determine an update fails we can
skip the update and continue to boot so it does not end up in a
reboot loop if the device has a persistent failure.

BUG=b:74336712
BRANCH=eve
TEST=manual: ensure update is only attempted if EC is updated

Change-Id: I71425b09c341bf0556612390f97329447da5eac4
Reviewed-on: https://chromium-review.googlesource.com/985395
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Duncan Laurie <dlaurie@google.com>
Tested-by: Duncan Laurie <dlaurie@google.com>
Trybot-Ready: Duncan Laurie <dlaurie@google.com>