Update patch set 1
Patch Set 1:

I'm still not really convinced about this one either, to be honest. This does effectively the same thing as the depthcharge patch we discussed: it causes errors during the aux FW update check to be silently ignored. And then we won't even attempt them again until the next ordinary firmware update.

I think the goal of this sync mechanism is to ensure that all parts of an update get applied correctly. If we can't do that due to whatever sort of error, we have to go into recovery mode. Otherwise nobody would ever notice these issues when they happen, especially if it's a rare failure.

Tying it to the EC update also seems problematic to me because the two updates can finish independently. It's perfectly possible that the EC update finishes but then the machine browns out or something like that during the following TCPC update (especially because they tend to be so damn slow). We should not silently continue with the old firmware in that case. (In fact, I don't think our TCPCs can even fall back to their old firmware when an update was interrupted, right? At least not all of them... e.g. the PS8751 doesn't have OTP, it just has a single reprogrammable SPI flash part, and if we end up with a broken, half-written piece of firmware on there it'll just not work at all. So it seems pretty important that we try these updates again until they stick, and use recovery mode to inform the user of repeated failures and avoid an update loop.)

Patch-set: 1
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>