Update patch set 6
Patch Set 6:

you should have the sys-devel/arc-cache-builder package install a program in `bin` that hides the `java ...` stuff.  that way we don't need the jar in this repo, and we can add something to the signing repo so it also internalizes the logic.

Patch-set: 6
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>