tpm_lite: tpmc command to check owner auth
Add a command that checks whether the well-known secret (all zeros) works for
owner authentication. This is accomplished by sending a DefineSpace command for
TPM_NV_INDEX_TRIAL, which will trigger auth checks but not actually allocate an
NVRAM space. The return value thus indicates whether authorization was
successful or not. This will be used in the tpm-firmware-updater driver script
to check whether the TPM is in upgradable state.

BRANCH=None
BUG=chromium:788719
TEST=compiles

Change-Id: I630831127e0e01186650412a92643c2153fbe2ee