Add support for booting into U-Boot
U-Boot is a widely used boot loader in the open source world. It provides
a wide array of boot options, including filesystems, many types of
peripherals, an environment for persistent state and a comprehensive
command line for debugging.

It is useful to be able to use U-Boot as the boot loader of choice when in
developer model. Add a way to support this, similar to the existing 'legacy'
boot-loader support.

BUG=chromium:837018
BRANCH=none
TEST=FEATURES=test emerge-grunt --nodeps vboot_reference

Change-Id: I02eab1b87e21a6401ec42317c4c1fa1bd2767b53
Signed-off-by: Simon Glass <sjg@chromium.org>