sign_official_build: switch kernel/firmware signing to loopdevs
Newer versions of util-linux/mount don't like when you create overlapping
loopback files.  Since we always create a loopback of the entire image,
this means every mount fails.

We can change the few users in here over to using the existing loopback
partitions rather than continuing to create their own from scratch.  This
makes the code a bit simpler.

However, we currently duplicate some of the mount image helpers so that
one version works off of a disk image while the other uses loopbacks.
Cleaning this up requires a number of changes in other files which we'll
want to do eventually, just not right now (to minimize risk).

BUG=chromium:714598
TEST=image signing works on newer gLinux installs
BRANCH=None

Change-Id: I31b35636b3b271e97070d283f8cb74d3183d8ec8